fetch github email data, verify email and username and login
Account should have unique email and unique username.
User is distinguished by their email so if user tries to social login and there is no existing email in userDB, that account would be a new user. But when its username is already exist, new username which is followed by single random number will be assigned.