change password
Since social-only user does not have password, they cannot access change-password page.
change-password-button would not be seen on edit-profile page and /users/change-password url would be also not accessible if it is social-only account.

like edit-profile, if password is updated, session should be also updated by assigning new password to original session-user data. But it can be improved by not using session data for password validation. Compare with password in database, not in session, and update it. updating session is not necessary.