fix deleteVideo related to comments
async/await does not work with forEach. use for in instead.
mongoDB has a version system, each document has its own version and version increases whenever document is re-saved. If the version does not match, save() does not work. In this case, commentOwner increases User's document version and user is also User's document, so user's version does not match with the value when it is assigned.